LU-9243 gss: fix GSS struct definition badness
struct lgssd_ioctl_param should not be defined in multiple places. So
move it to a new header file lgss.h that can be included from kernel
space and user space.

struct gss_header, struct gss_rep_header, struct gss_err_header and
struct gss_wire_ctx are going on the wire, so they need to be moved to
lustre_idl.h, and be wire-checked.

Reviewed-on: https://review.whamcloud.com/46543
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>

+++ lustre/utils/gss/lgss_keyring.c
@@ -52,6 +52,7 @@
 
 #include <libcfs/util/param.h>
 #include <libcfs/util/string.h>
+#include <uapi/linux/lustre/lgss.h>
 #include "lsupport.h"
 #include "lgss_utils.h"
 #include "write_bytes.h"
@@ -260,7 +261,7 @@ int do_nego_rpc(struct lgss_nego_data *lnd,
 	logmsg(LL_TRACE, "do_nego_rpc: to parse reply\n");
 	if (param.status) {
 		logmsg(LL_ERR, "status: %ld (%s)\n",
-		       param.status, strerror((int)(-param.status)));
+		       (long int)param.status, strerror((int)(-param.status)));
 		return param.status;
 	}
